Update patch set 2
Patch Set 2:

In addition to true/false vs. on/off - Samsung seems to have swapped the meaning of true/false on Exynos devices.

On Exynos devices, dualmic_enabled=false causes the NC mic to be enabled.  dualmic_enabled=true causes the NC mic to be DISABLED.  On Exynos devices, the best way to test is to speak into the top of the phone (e.g. speak into the speaker) which is where the NC mic is located - If the other end can hear you, the NC mic is on.

Obviously on your hardware since you're building from source you can set this properly.

Patch-set: 2